Create escutcheon
Collect a dataset of audio samples or MIDI files that represent the sounds and musical sequences that you want the synthesizer to be able to generate.

Preprocess the data to prepare it for training. This may involve normalizing the data, selecting a subset of the data to use for training, or converting the data into a format that is suitable for machine learning.

Train a machine learning model on the dataset. This could be done using a variety of machine learning algorithms, such as a neural network or a decision tree. The specific algorithm and hyperparameters used will depend on the characteristics of the data and the desired output of the synthesizer.

Implement the synthesizer as a software program that can generate audio or MIDI data based on the trained model. This may involve building a user interface and integrating the machine learning model into the synthesizer software.

Test and refine the synthesizer to ensure that it is producing the desired output. This may involve adjusting the machine learning model, the synthesizer software, or both.